remove build-dist.js
